Connection Flow and Object
Created a CumulusCI project.  Grabbed the Flow created during the Feb. Sprint and the Custom Object.

Unable to launch a scratch org because of a PMM dependency within the Flow.  Need to add PMM to the install along with NPSP Dependency, so V4S and PMM are both a part of the install.