Fixed Flow Deploy issue
The Flow wasn't able to deploy because it needed to reference the Connection field on the PMM and V4S objects.  Added those fields to the repo, so when it spins up, all is deployed.  Flow is Active.

Just need to add Connection fields and objects to the profile to see them appropriately.